Update to support inverse value for refunds
Due to a bug in the Fat Zebra sandbox the refund amount was being sent through as a negative amount, whereas the Live gateway would return a positive amount. This means that in a live website the refund was being counted as a positive amount and skewing the order transactions.

This change checks for the '-R-' in the ID (which indicates a refund record) and flips the value (e.g. multiply by -1 to make it a negative) before applying it to the transaction.